+++ src/lib/locales/en.json
@@ -87,6 +87,7 @@
     "Notifications": "Notifications",
     "One-time": "One-time",
     "Ongoing": "Ongoing",
+    "Open events page": "Open events page",
     "Operational": "Operational",
     "Partial Degraded Performance": "Partial Degraded Performance",
     "Partial System Outage": "Partial System Outage",
