+++ src/routes/(kener)/monitors/[monitor_tag]/+page.svelte
@@ -43,7 +43,11 @@
   {/if}
 </svelte:head>
 <div class="flex flex-col gap-3">
-  <ThemePlus monitor_tags={[data.monitorTag]} embedMonitorTag={data.monitorTag} />
+  <ThemePlus
+    monitor_tags={[data.monitorTag]}
+    embedMonitorTag={data.monitorTag}
+    hideNotificationsPopover={showInlineEvents}
+  />
   <div class="flex flex-col gap-2 px-4 py-2">
     {#if data.monitorImage}
       <img
