+++ src/routes/(manage)/manage/app/site-configurations/+page.svelte
@@ -510,7 +510,8 @@
     const defaults = structuredClone(defaultEventDisplaySettings);
 
     return {
-      showInlineEvents: Boolean(parsed?.showInlineEvents ?? defaults.showInlineEvents),
+      showInlineEvents:
+        typeof parsed?.showInlineEvents === "boolean" ? parsed.showInlineEvents : defaults.showInlineEvents,
       incidents: {
         ...defaults.incidents,
         ...parsed?.incidents,
